Move deface override file to method level file
